Move 'FormErrors' to common
